Included Switch Camera to First Person
feat: Able to switch between First Person and Third Person

The turning of corners is working as good as usual. Changes all occured
in RunCharacter. Created a custom function "SwitchViewpoint" and Boolean
"IsInFPCamera". In the Viewport of RunCharacter, I added a FPCamera
which is attached to the Mesh and what is special about it is "Use Pawn
Control Rotation" is enabled in this camera which the "FollowCamera"
does not.