feat: Introduced death animations and fixed a few bugs
Followed the Unreal Twin-stick tutorial death animations part to learn
how to set the death animations up. There are 3 animations that will be
played randomly too to add some uniqueness. In addition, I think we
should make some wise-ass cracks jokes when dying to make the death
funny. In a way to entice people to die, though still keep running in
the game.