feat: Introduced Gamepad support
Now you can use Gamepad to play the game, with the left thumbstick to
control the player movement and right thumbstick to control the torch
light shining. I have also got the turn corners working by using the
left thumbstick, Hit left with the thumbstick to tun left, Hit right
with the thumbstick to turn right.